écrans de chargement niveaux etc....

+++ src/com/bonhomi/game/GameManager.java
@@ -1,21 +1,20 @@
 package com.bonhomi.game;
 
 import java.awt.Color;
+import java.awt.Font;
 import java.awt.Graphics2D;
 import java.awt.Point;
 import java.awt.Rectangle;
-import java.awt.Shape;
-import java.awt.geom.GeneralPath;
-import java.awt.geom.Rectangle2D;
-import java.awt.geom.RoundRectangle2D;
+import java.awt.image.BufferedImage;
 import java.util.Timer;
 import java.util.TimerTask;
 
+import javax.swing.ImageIcon;
+
+import com.bonhomi.main.Afficheur;
 import com.bonhomi.main.Core;
 import com.bonhomi.main.Loopable;
-import com.bonhomi.main.MainClass;
 import com.bonhomi.main.SpriteLoader;
-import com.bonhomi.main.SpriteOccurence;
 import com.bonhomi.sounds.SoundSystemMaster;
 
 
@@ -27,48 +26,27 @@ public class GameManager implements Loopable {
 
 	static private boolean initialized = false;
 	static final Rectangle window = new Rectangle(0, 0, Core.WIDTH, Core.HEIGHT);
+	
+	static private final BufferedImage loadingImg = (new SpriteLoader("UI/game/", "Loading")
+	).getActualImage();
+	static private final BufferedImage gamefinishedImg = (new SpriteLoader("UI/game/", "gameFinished")
+			).getActualImage();
+	static private final BufferedImage gameOverImg = (new SpriteLoader("UI/game/", "gameOver")
+			).getActualImage();
+	
 
 	static Player player1;
 	static BadGuys[] ennemis;
 	static NavMesh nav_mesh;
 	
 	static private Timer damageTimer;
+
 	
 	static GameUI GUI;
 	
-	//création d'un nav_mesh pour le niveau
-	private static final int offsetMurs = 100;
-	private static final int widthPortes = 75;
-	private static final Shape[] compo_navigation = {
-			//salle 0
-			new Rectangle(offsetMurs, offsetMurs, Core.WIDTH-2*offsetMurs, Core.HEIGHT-2*offsetMurs)
-			.getBounds2D(),
-			
-			//porte haute 1
-			new Rectangle(Core.WIDTH/2 - widthPortes/2, 0, widthPortes, offsetMurs)
-			.getBounds2D(),
-			//porte basse 2
-			new Rectangle(Core.WIDTH/2 - widthPortes/2, Core.HEIGHT-offsetMurs, widthPortes, offsetMurs)
-			.getBounds2D(),
-			
-			//porte gauche 3
-			new Rectangle(0, Core.HEIGHT/2 - widthPortes/2, offsetMurs, widthPortes)
-			.getBounds2D(),
-			//porte droite 4
-			new Rectangle(Core.WIDTH-offsetMurs, Core.HEIGHT/2 - widthPortes/2, offsetMurs, widthPortes)
-			.getBounds2D(),
-	};
-		
-	//test obstacle
-	/*private static final Shape[] compo_obstacle = {
-			//pillier central 0
-			new RoundRectangle2D.Double(
-					Core.WIDTH/2 - widthPortes, Core.HEIGHT/2 - widthPortes, 
-					widthPortes*2, widthPortes*2, 
-					widthPortes*2, widthPortes*2)
-	};*/
+	static protected Niveau niveau1;
 	
-	protected Niveau niveau1;
+	private float time;
 	
 	// Constructeur
 	public GameManager()
@@ -76,7 +54,7 @@ public class GameManager implements Loopable {
 		GUI = new GameUI();
 		damageTimer = new Timer("GestionDegats");
 		nav_mesh = new NavMesh();
-		this.niveau1 = new Niveau();
+		niveau1 = new Niveau();
 		
 		init();
 	}
@@ -94,12 +72,19 @@ public class GameManager implements Loopable {
 				@Override
 				public void run() 
 				{
-					for (BadGuys ennemi : ennemis)
+					if (player1.getVie() <= 0)
+						this.cancel();
+					
+					Entity[] eList = niveau1.getActualRoom().getEntityList();
+					for (Entity e : eList)
 					{
 						/* on retire des point de vie au joueur si il touche un
 						 * ennmis et qu'il lui reste des vies
 						 */
-						if (player1.intersects(ennemi) && (player1.getVie() > 0))
+						if(!e.isEnnemy())
+							continue;
+							
+						if (player1.intersects(e))
 						{
 							player1.perdreVie();
 							SoundSystemMaster.getInstance().ouille();
@@ -115,60 +100,70 @@ public class GameManager implements Loopable {
 	@Override
 	public void init()
 	{
-		player1 = new Player(200, 300, 1);
-		
-		//initialisations...
-		player1.init();
-		GUI.init();
-		damageTimer.purge();
-		nav_mesh.purge();
-		
-		niveau1.printMap();
-
-		//on ajoute le tout au nav_mesh
-		nav_mesh.addNav(compo_navigation);
-		//nav_mesh.addObs(compo_obstacle);
-		
-		//on selectionne le joueur 1 en tant qu'entité a surveiller avec l'UI
-		GUI.setPlayerFocus(player1);
-		
-		//on cree les ennemis sur le terrain
-		ennemis = new BadGuys[2];
-		ennemis[0] = new BadGuys(200, 110, 1);
-		ennemis[1] = new BadGuys(400, 110, 1);
-		
-		for(BadGuys b_g : ennemis)
+		if (niveau1.isLoaded())
 		{
-			b_g.init();
+			player1 = new Player(200, 300, 1);
+			
+			//initialisations...
+			player1.init();
+			GUI.init();
+			
+			damageTimer.purge();
+			nav_mesh.purge();
+			
+			niveau1.init();
+			
+			//on selectionne le joueur 1 en tant qu'entité a surveiller avec l'UI
+			GUI.setPlayerFocus(player1);
+			
+			//on lance le fil d'execution gerant les degats:
+			playerDamage();
+			
+			initialized = true;
 		}
-		
-		//on lance le fil d'execution gerant les degats:
-		playerDamage();
-		
-		initialized = true;
+		else
+			niveau1.startLoading();
 	}
 
 	// Affichage
 	@Override
 	public void draw(Graphics2D g) 
 	{
+		//on dessine tout!!
+		if (!niveau1.isLoaded())
+		{
+			//dessin de l'ecran de chargement
+			g.drawImage(loadingImg, 0, 0, null);
+			
+			//on enregistre l'ancienne fonte pour la replacer après
+			final Font recup_font = g.getFont();
+			
+			//nouvelle fonte à afficher
+			final int fontSize = 100;
+			final Font font = new Font("Monospaced", Font.BOLD, fontSize);
+			g.setFont(font);
+			
+			//clignotement du texte chargement avec la fonction Math.cos(a)
+			g.setColor(new Color( 0.95f, 0.95f, 0.5f, (float) ( Math.abs(Math.sin(time)))));
+			if( time >= Math.PI*2)
+				time = 0;
+			else
+				time += Math.PI/75;
+			
+			//dessin du texte
+			g.drawString("Chargement", (int) (Core.WIDTH/2 - fontSize*3.11), Core.HEIGHT/2);
+			
+			g.setFont(recup_font);
+			
+			return;
+		}
+		
 		g.setColor(Color.white);
 		niveau1.draw(g);
 		
-		/*if( MainClass.getDebugLvl() > 2)
-		{
-			g.setColor(Color.yellow);
-			g.fill(nav_mesh);
-			g.setColor(Color.orange);
-			g.draw(nav_mesh);
-
-		}*/
+		//un dessin ne ressort de là que si le debug est activé
+		nav_mesh.draw(g);
 		
-		g.setColor(Color.red);
-		for(BadGuys b_g : ennemis)
-		{
-			b_g.draw(g);
-		}
 		
 		g.setColor(Color.green);
 		player1.draw(g);
@@ -183,47 +178,36 @@ public class GameManager implements Loopable {
 	 */
 	public void update() 
 	{
+		if (!niveau1.isLoaded())
+			return;
+		if (niveau1.isLoaded() && !initialized)
+			init();
+
+		
 		if(!initialized) 
 			throw new IllegalStateException("Class Updated before Init!");
+		//on nettoie le nav_mesh pour les entités dynamiques
+		nav_mesh.purge();
 		
+		//le niveau régénère le nav_mesh par la methode statique
 		niveau1.update();
 		
+		
 		player1.update();
-
-		//on purge le nav_mesh ici pour que les ennemis puisse toucher le joueur.
-		nav_mesh.purge();
-		//on ajoute le tout au nav_mesh et la case obstacle du joueur
-		nav_mesh.addNav(compo_navigation);
-		//nav_mesh.addObs(compo_obstacle);
 		nav_mesh.addObs(player1.ObsComp());
 		
 		GUI.update();
 		
-		for(BadGuys b_g : ennemis)
-		{
-			//les ennemis sont dans la fenetre de jeu avant traitement
-			if (window.contains(NavMesh.getNavPoint(b_g)))
-			{
-				//on fait en sorte que les ennemis ne se supperposent pas
-				nav_mesh.addObs(b_g.ObsComp());
-				
-				Point suivre_joueur = new Point((int) player1.getCenterX(), (int) player1.getCenterY());
-				
-				if (b_g.Cible != null)
-					b_g.Cible.setLocation(suivre_joueur);
-				else
-					b_g.Cible = suivre_joueur;
-				
-				b_g.update();
-			}
-			else
-				b_g.Cible = null;
-		}
+
 	}
 
 	@Override
 	public void terminate() 
 	{
+		if(!initialized)
+			return;
+		
+		niveau1.terminate();
 		GUI.terminate();
 		player1.terminate();
